fakeroot is graceful on second run 	modified:   board/d-tacq/acq400/fakeroot.sh

+++ board/d-tacq/acq400/fakeroot.sh
@@ -2,5 +2,11 @@
 
 
 echo FAKEROOT01 promote libtirpc to /lib .. needed by busybox
-mv output/target/usr/lib/libtirpc.so* output/target/lib
+if [ -e output/target/lib/libtirpc.so.3.0.0 ]; then
+	echo FAKEROOT45 job already done
+	ls -l output/target/lib/libtirpc.so*
+else
+	mv output/target/usr/lib/libtirpc.so* output/target/lib
+fi
+
 echo FAKEROOT99
